+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/system/DeriveController.java
@@ -10,6 +10,7 @@ import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
 import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
 import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -50,6 +51,7 @@ public class DeriveController {
     @PostMapping("/derive/dynamic-derive")
     @BehaviorAnalyse
     @BindingCheck
+    @LoginRequired
     public FastJsonResponseData<JSFixedFastJsonDeriveResponse> dynamicDerive(
             HttpServletRequest request,
             @RequestBody @Validated WebInputDynamicDeriveInfo deriveInfo, BindingResult bindingResult
@@ -69,6 +71,7 @@ public class DeriveController {
     @PostMapping("/derive/static-derive")
     @BehaviorAnalyse
     @BindingCheck
+    @LoginRequired
     public FastJsonResponseData<JSFixedFastJsonDeriveResponse> staticDerive(
             HttpServletRequest request,
             @RequestBody @Validated WebInputStaticDeriveInfo deriveInfo, BindingResult bindingResult